+++ tests/utils.py
@@ -643,6 +643,8 @@ def execute(command, data_in=None, timeout=0, error=None):
     @param command: executable to run and arguments to use
     @type command: list of unicode
     """
+    if PY2:
+        command.insert(1, '-W ignore::FutureWarning:pywikibot:132')
     if cryptography_version and cryptography_version < [1, 3, 4]:
         command.insert(1, '-W ignore:Old version of cryptography:Warning')
     # Any environment variables added on Windows must be of type